Add inline operator support + stats implementation
Summary:
Inline operators are operators that process tuples value by value without using a splitter (which requires materialization).
They can be inserted inline into a tuplestream with an InlinedOperatorHook. Wrapping an inlined operator with FullInlinedOperatorTupleStream
or PartialInlinedOperatorTupleStream gives the view of it as a tuplestream that produces some values.

In the stats case, we can produce partial results by querying the results before the full stream has been processed.