Add SimpleSqlExecution
Allows early peel-out of simple SQL queries that do not need to go through the full planning
and can be executed on the coordinator directly. This is not the end stage (we want to run them
through the analyzer and react there) but for now this is good enough to execute commands like
IMPORT, DROP TABLE etc.