Simple heartbeat-based failure detector
This is a naive failure detector that watches discovery for "presto" announcements and monitors the remote services using the http url.

It considers a service failed if the percentage of failed requests over the last minute (exponentially decayed) is higher than a configurable threshold. It also supports a configurable warmup period during which newly discovered nodes are considered failed.