allow empty address lists in native split
NativeSplit only needs an address when reading data. When data is
written, the node that will hold the data will be known only after the
data was written (and the split was consumed). This was masked till
now by using the list of addresses from the source split (the hdfs
nodes that hold the source data) but the raid changes now may return
empty lists.

Fix is to allow empty lists, set the list explicitly to empty when
creating a write split and check to make sure that the list of
addresses returned in the native split manager is not empty.