Add session to ConnectorMetadata calls
This is an alternate implementation of pull request #1159  The main difference
being this code add Session to each metadata function not involving a handle,
instead of making ConnectorMetadata a stateful instance. The session is not
passed to handle based methods and instead any connector needing session data
would need to store the session in the handles. This is consistent with the
handle design of other methods.