Require terminated statements from files in CLI
Previously, the CLI would automatically terminate the last statement
read from a file. A non-terminated statement will now result in an
error. This is because there is no way to detect if the partial
statement is intended or if it is the result of an error condition
such as a broken pipe.

For example, consider the following statement:

    SELECT * FROM t
    WHERE id = ...;

Suppose an external program was feeding this statement to the CLI
using a pipe that was broken after the first line. This would result
in the following statement being executed:

    SELECT * FROM t
    ;

Such a semantic difference could have disastrous consequences,
especially when combined with data modification commands (which we
plan to support in the future).