Don't include plugins as presto-server dependencies
Dependency resolution rules in the maven assembly plugin are broken.
This is what seems to be happening. Consider the following dependency tree:

presto-server -> commons-lang
              -> cassandra plugin -> commons-lang

The assembly exclusion rules say "don't include cassandra plugin and its transitive dependencies".
Unfortunately, the rule applies to commons-lang and takes precedence over the natural inclusion
of that dependency by presto-server and it's not added to the tgz.