Add ByteCodeExpression abstraction
A ByteCodeExpression is chain of Java like expressions that results in a single
value being pushed on the stack.  The chain starts with a constant, local
variable, static field, static method or invoke dynamic followed followed by
zero or more invocations, field dereferences, array element fetches, or casts.

A ByteCodeExpression is a ByteCodeNode so it works with tools like tree dump.

This abstraction makes it easy to write generic byte code generators that can
work with data that may come from a parameter, field or the result of a method
invoation.