Update provisio for correct ordering for parallel builds.
The graph edges were being created using GAC, but inside the reactor they need to be
created using GA in order to affect the reactor sorting (correct toposorting). With this
change projects can be built in parallel, correctly.

I have also turned off the dependency checking as the dependencies are synthesized from
looking at the provisioning descriptor and will be analyzed and deemed extraneous. None
of the dependencies will link to any classes because this project is purely an assembly
so what is required is for the provisioning descriptor to decide.