change array, map and row type encoding
Current Array, Map and Row type encoding which uses JSON has a limitation
that not null string can be used as Map keys.
And even though JSON is very user and developer friendly,
there would be a higher serialization/deserialization cost and doesn't seems to cooperate with other primitive types.

As an alternative encoding, we can simple think Array as the VariableWidthBlock encoded Slice.
So to speak, Array ["a","b","c"] is a VariableWidthBlock having 3 positions and can be easily serialized as Slice.
Map also can be though as an Array of [k1, v2, k2, v2, ...] and Row as [field1, field2, ...]

As different type can have different machasism to be storead in a BlockBuilder. For example, Date types serialize its value in 4 bytes, ArrayBuild should aware exact matching parameter types based on its struct type.