Add support for AWS EC2 IAM Roles
This update adds support for EC2 Instances that are launched with IAM Roles. 
The change will determine whether the credentials are configured in the xml and if not present assume the credentials provided by the IAM Role associated with the EC2 Instance. 
More information on IAM roles here: http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/iam-roles-for-amazon-ec2.html