Handle ORC files with corrupt checkpoints
When ORC switched to low memory mode it can create invalid checkpoints.  When
reading the file if the checkpoints can not be decoded, treat the entire strip
as a single row group.  DWRF files with a row group dictionary still fail
because the row group dictionary length is contained in the checkpoint stream.