Introduce TableLayout abstraction
This commit introduces the concept of a TableLayout, which represents a physical
organization of a Table. A table can have multiple physical layouts with different
properties (partitioning, grouping, predicate, columns, etc).

Presto will be able to take advantage of this information to pick the layout
most suitable to satisfy a given query plan based on desired properties.

Partition/ConnectorPartition/etc. are now deprecated, but there's an adaptation
layer that allows existing connectors to work seamlessly with the new APIs. Once
all connectors are migrated, those APIs will be removed.