Enforce JOINs to have the same output symbols after PredicatePushDown
After PredicatePushDown optimizer, JOINs may produce additional output symbols if the JOIN clause was rewritten. This impacts nodes such as ExchangeNode which require that the inputs from each underlying source matches the output.
In this diff, we enforce that JOINs always have the same output symbols after PredicatePushDown. We also add some checks to ExchangeNode to catch when the inputs do not match the outputs.