Fix race freeing memory
The state change listener in TaskContext races against
Driver.destroyIfNecessary to free the memory held by the task. Since the
state change listener is only there as a safety check, remove it.
Also move the freeing of the memory into the destroy operator loop, so
that the memory is freed from the OperatorContext as well.