Fix query/stage final state race condition
Currently, the query listens for output stage state changes and each stage
listens for stage changes from the children.  When a stage finishes, the
parent and children are aborted, and the query sees an aborted output stage.
Then the query searches for the failure info.  This creates race conditions
resolving the final state.

This change avoids these problems by not having stages listen for state changes
from substages, and instead have the query listen for state changes on all
tasks.  If a task fails, the query moves to the failed state first, and then
all stages are aborted.  This top down approach eliminates the races since the
query always records the final state before cleanup.