Fix leak in SqlStageExecution
Task scheduling and stage cancellation race against each other. In
particular, scheduleSourcePartitionedNodes() may take a long time if it
has to wait for free nodes, and a LIMIT query could finish after just
one of those tasks starts running. In this case the stage will be
cancelled while scheduling is still in progress, which will cause tasks
to be leaked, as they're created after the stage already has
transitioned to the CANCELLED state.