Build RPM for presto server
Added the rpm build logic to a new module called presto-server-rpm.
We use the tar.gz built from presto-server module, untar it to
presto-server-rpm/target and then use it to build the rpm with rpm-maven-plugin

The rpm build is optional and will only be invoked using '-P rpmbuild'.
Run: 'mvn clean install -P rpmbuild' and rpm will be under
presto-server-rpm/target/rpm/presto/RPMS/x86_64

In post-install of rpm, we want to populate the node.properties file originally
templated with the node.id to have the actual value of uuidgen on the
node. This will enable us to have a single-node presto installed by just
running rpm -i.

In pre-install, we check if Java installed is Oracle's java version 1.8.
The install will fail if this vendor and version is not found. We did
not want to add jdk as the presto rpm pre-requisite, hence this check.

Also added init.d scripts to run start/stop/restart and status
on Presto server. We also provide chkconfig to init.d script.