Fix aggregation test utilities
Currently in partialAggregation, all page are aggregated to produce one
intermediate result. And it gets combined with an empty intermediate result to
produce the final result.  Instead, it should produce one intermediate result
for each page.

This bug gives false confidence that combine() is correctly implemented for an
aggregation function. When in fact, combine() is hardly tested.

assertAggregation is also fixed to produce 2 pages when possible.