connector: Add a connector for Phoenix
Phoenix provides a SQL based access to HBase. This implementation
is for HBase version 0.98.13 and pheonix-4.4.0-HBase-0.98.

Minor change was done in QueryBuilder to use constant 1 instead
of null when columns are not present in the query. Current Phoenix
thin client has a bug with SELET null queries. Using 1 instead of
null will not change any functionality for Presto.

Manual testing has been done, product tests will be added later.