bugfix - ColumnJdbcTable was trying to set an int into a column which expected bigint
This bug intermittently was causing TestDriver tests to fail.
ColumnJdbcTable was trying to set a result from
com.facebook.presto.connector.system.jdbc.ColumnJdbcTable#jdbcDataType
which is an int for a 'data_type' column which is a BIGINT.

And the same happened for all the BIGINT columns.

I don't know why it happens intermittently.

Test Plan: running unit tests