Insert into a table with a list of columns
Now it is possible to specify list of columns with INSERT INTO e.g.
INSERT INTO orders (orderkey) VALUES(1);

When the list of column names is specified the columns produced by the
query must exactly match to them. Each column not being present in
the column list will be filled with a ``null`` value.
In case the list of column names is not given then the columns produced
by query must exactly match the columns in the table being inserted
into.

When the list of columns is present then there is added additional
ProjectNode to logical plan which represents mapping between query
columns and table columns. Also it takes care of table columns not
present in a list.