Do not escape schemaName in MySqlClient#getTables()
In MySqlClient#getTables() schemaName is passed as catalog parameter of
DatabaseMetaData#getTables(). It is because MySql do not have schemas.
DatabaseMetaData#getTables() requires that catalog parameter to match
catalog name. There should be no wildcards escaping there, as it does
not accept pattern as it does for tableName or schemaName.

Fixes #3763.