Fix handling of MySQL databases with underscores
In MySqlClient#getTables() schemaName is passed as catalog parameter
of DatabaseMetaData#getTables() becasue MySQL does not have schemas.
DatabaseMetaData#getTables() requires the catalog parameter to match
the catalog name. There should be no wildcard escaping there, as it
does not accept a pattern as it does for tableName or schemaName.