Add the Maven Wrapper to Presto
This will allow Presto to transparently use newer versions of Maven, and transparently allow
users to take advantage of those features without any burden on their part.

This will not affect any current users and this feature can bake as the
documentation is updated on the Presto site and is seen by users. I
recommend leaving the wrapper in the build for a few weeks before
adding any features that require its use.

More information about the wrapper can be found here:

https://github.com/takari/maven-wrapper