Convert type to hive type when adding a new column
When a new column is added to an existing table with an `alter table`
statement the type of the new column is not converted to an appropriate
Hive type unlike the `create table` statement. For example,

create table testvarchar (name varchar);

This `create table` statement creates a hive table with a name column of type string.

When we do an alter table:

alter table testvarchar add column last_name varchar;

the new column last_name is added to the testvarchar table as a varchar
type instead of a string type causing the subsequent queries to the
testvarchar table to fail.