Update redline rpm plugin version
Update the version fo redlinerpm-maven-plugin to the latest.  This fixes
the error that occured if a build was attempted when artifacts existed.
Now it will always overwrite existing artifacts.

This version of redlinerpm-maven-plugin also brings in the latest version
of redline-rpm, which has some fixes for compatiblity with later versions
of the rpm library.  One of those fixes- supporting large inode values -
is not compatible with versions of rpm before 4.6.
Add a profile that builds the rpm using the older version of redline-rpm

Testing: confirmed could build multiple times. Confirmed inode type in
rpm header is 32bit int for sles profile and 64bit int for the default.