Use task intsance ID instead of node instance ID
When a client tries to talk to a task to a non-exisiting task, we create
a new task. In some cases it recreates tasks that were previously torn
down. Attach a task instance id to each task so that the client can
identify that the task it was communicating with has changed.