Fix invalid bucket calculation HashPartitionMaskOperator
Math.abs returns a negative value for MIN_INT, so when this hash value
is later modded to choose a bucket, the bucket is also negative.  In the
HashPartitionMaskOperator this means the value is a member of no buckets
and the value is skipped.

The HashPartitionMaskOperator is only used for the experimental local
parallel aggregations optimization so this not a high impact bug.