Require that clients explicitly announce transaction support
Pre-transaction clients will not prevent a user from sending a BEGIN statement
and will fail to follow the new transaction protocol. This means that a user
could run a successful BEGIN statement, and the following statements would be
executed in an autocommit mode because the client is not supplying the
transaction info.