Fix sporadic terminal reset issue in CLI
After the user exits the pager, the output handler may still be in a loop
waiting for data to write to the output printer. If so, we need to
interrupt the thread in order to immediately return control to the user.
However, we don't want to interrupt if it has already returned and is
doing something else, such as resetting the terminal settings.

This fixes the "Failed to disable interrupt character" error.