Remove usages of canHandle
Handles are polymorphic. When deserializing, Jackson needs to know the
concrete type to deserialize, but it only has the supertype available
and thus needs help to determine the concrete type. Jackson supports
this by additional type identifier to the serialized object.

The previous approach was to use the connector ID as the type identifier
and to determine the ID when serializing by asking each connector if a
given handle instance belongs to that connector by calling canHandle().
This requires writing a lot of boilerplate for every connector and to
store the ID in each handle when it would not otherwise be needed.

The new approach is to use the connector name rather than the connector ID
for the type identifier and to use the class identity of the handle to
determine the connector to which it belongs. This introduces the
requirement that connectors use the same handle classes for all instances
of the connector.

This commit only changes serialization to use class identity. The other
changes are done in a future commit.