Avoid unnecessary encoding/decoding when writing VARCHAR columns
The current workflow turns VARCHAR columns from Slice to String to
Text (i.e. back to bytes) before serializing them.

This change avoid that unnecessary decoding/encoding step by feeding
Hive with Text objects directly.

It also includes an optimized version of Hive's columnar SerDe that
avoid a call to get the String value in serialize(), which is only
used to check whether the string is empty.

A test with a table containing 13 varchar columns showed a reduction
in CPU utilization of ~28%.
An informal benchmark showed an improvement in rows / CPU second of 37%
when writing a table with 13 VARCHAR columns.