Optimize writing Hive tables
- Avoid unnessary encoding/decoding when writing VARCHAR columns

Currently, the flow in HivePageSink converts VARCHAR columns from
Slice to String to Text before they get written. We now bypass the
conversion to String by dealing directly with Text objects.

- Avoid unnecessary type equality checks to decide how values should be
  extracted from column Blocks.

Instead, build field "setters" and rely on the JVM to do the dispatch.
Also, avoid creating boxed objects for each value that's written.

A test with a table containing 13 varchar and 3 bigint columns showed
a reduction in CPU utilization of ~34%.