Introduce Apply plan node
It is a logical representation of a subquery.
Apply node contains input relation, expression relation and parameters
(symbol list).

Each row in the input relation is passed to the expression, then
expression (subquery) is executed. Expression produces new columns which
are joined with the input row. Parameters (correlation list) determines
what columns need to be passed from the input to the expression to
calculate the result rows. If correlation list is empty then subquery
is not correlated, in this case values returned from expression are
always the same, not dependent on input relation.

See
https://docs.google.com/document/d/18HN7peS2eR8lZsErqcmnoWyMEPb6p4OQeidH1JP_EkA/