Resources should be closed
Java's garbage collection cannot be relied on to clean up everything.
Specifically, connections, streams, files and other classes that
implement the Closeable interface or it's super-interface,
AutoCloseable, must be manually closed after creation. Failure to do so
will result in a resource leak which could bring first the application
and then perhaps the box it's on to their knees.