Use an old testng for all but presto-product-tests
Somewhere between 6.8.21 and 6.9.6, an issue is introduced where the
@BeforeClass methods in children of AbstractTestHiveClient don't always get
called.

Originally, presto-hive-* use the older version, while others use the newer
version. However, this causes dependency conflicts in upcoming changes. To
resolve the conflict, presto-tests must use the old version. And every
connector depends on presto-tests. As a result, most of Presto must use the old
version.

This commit changes the version of testng in all Presto artifacts expect
presto-product-tests to the old version.