Ensure all DELETE requests in HttpPageBufferClient are received
If ExchangeClient.close() is called a DELETE is sent without retry. This
means that the message may get lost, causing the SharedBuffer on the
other side to not be cleaned up until it becomes abandoned after the
entire query finishes. This could also theoretically cause a deadlock,
as the SharedBuffer won't purge pages for this client. This fixes these
problems by waiting until the DELETE is received, and failing the query
if it is not received.