Support uncorrelated subqueries in aggregation queries
Until now all the subqueries were converted (to some form of join)
before applying filtering, aggregation, window or order by
clauses, independently on where the subquery was placed.

To enable subqueries for queries with aggregation above behaviour had to
be changed. Depending on where the subquery is placed, it had to be
applied at different processing moment in logical planner.

Subquery in filter and group by clause is applied before aggregation.
Subquery in having, project and order by clause is applied after aggregation.

Hovewer if the same subquery expression is used in project, group by
and/or having clause, then subquery have to be applied before
aggregation and places where the subquery was used are replaced by the
symbol of the subquery application.