Add validation for partition key ordering in Hive
Hive requires that partition keys be the last columns in the table and
that they be specified in the same order as they are in the table
properties. Added validation to check that the CREATE TABLE statements
with Hive targets contain columns in the right order. Previously, the
columns would appear to be arbitrarily re-ordered, which could cause
confusion.

Also fixed tests that specified columns in the wrong order.