Fix broken assertion in DomainTranslator
With the introduction of parametric varchar, the constraint that types must
be equal no longer holds.

The updated code checks whether the value is coercible to the field type
and then delegates to the registered coercion to transform it.

We introduce a FunctionInvoker that encapsulates the logic required to call
registered functions (binding session argument, etc).

This also fixes a NPE introduced with the parametric varchar change when trying
to construct a NullableValue if the value is null.