Allow type-compatible changes to view definitions
If the output of a view changes in a type-compatible way (e.g.,
due to changes to underlying tables or inferred expression types), don't
consider it stale.

A type-compatible change is any change where the actual type of running
the view query is implicitly coercible to the declared type of the view.

The planner now adds coercions if necessary to produce the expected output.