Don't use Count(*) to estimate partition columns in Cassandra >= 2.2
Previously, the Cassandra connector would do a SELECT * FROM x WHERE pk IN (a,b)
if the number of partition keys were less than 200. In Cassandra < 2.2, the query
SELECT COUNT(*) FROM x LIMIT 200 would return at most 200 if there were more than
200 rows.

In Cassandra 2.2 (CASSANDRA-8216) this behavior was changed, and now COUNT(*)
accurately returns all the rows in the table. Not only is this slow, but the
logic in the code also assumed that if the number of rows returned was anything
but 200 it would only query (up to) 200 partitions (Which is what caused #4014).

This commit simply disables this optimization for Cassandra 2.2+. In the future,
for table scans Presto should also move away from the thrift API and move towards
 the size_estimates which should reduce needs to optimizations like this.