Introduce most specific function selection algorithm
If more that one function is selected as applicable
for some particular parameters, the latest one in the candidates list
is being selected for execution by analyzer. Candidates list is built with reflection,
so the candidates list order is runtime specific.

This commit introduces most specific function selection algorithm.
This algorithm is suppose to ensure that selected function for some particular
parameters will always be the same, and it will always be the most specific.

The main algorithm idea is inspired by Java language specifications:
https://docs.oracle.com/javase/specs/jls/se8/html/jls-15.html#jls-15.12.2.5

However the actual implemented algorithm has a several differences.
Implemented algorithm in Presto is always trying to select applicable function
that doesn't require any coercions for invocation, no matter what type of function
it is - parametrized, variable arity or generic.

The actual implemented algorithm is next:

1.	Try to select function with explicitly declared parameters by direct match
2. 	If more than one function selected than throw "ambiguous method invocation exception"
3. 	Try to select generic function without coercion
4. 	If more than one function selected than throw "ambiguous method invocation exception"
5. 	Select all applicable function with coercions
6. 	Select most specific function
   	One function is more specific than another if the invocation handled by the first method
   	could be passed on to the other one
7. 	If the most specific function can't be selected than throw
   	"ambiguous method invocation exception"

Although this algorithm covers most of the regular cases, some cases for `UNKNOWN` parameters
are covered as a special case. `UNKNOWN` actually is not a real type, but rather some specific
type for the values, types of which is unknown. So the coercion from `UNKNOWN` to Type can
be considered as type inferring based on the context.

So in addition to the above 7 steps, when one of the arguments is unknown, the following
3 checks are being applied before throwing the `ambiguous call` exception.

1.	Try to select single function that require only `UNKNOWN` to `TYPE` coercion.
	If single function is selected - return it as a result.
2.	If the return type of the multiple selected functions is the same, and
	the function parameters are not declared as nullable, all that functions
	are being considered as semantically equivalent, and very first of those
	is being returned as a selected function.
3.	If still more that one function are applicable than throw `ambiguous call`
	exception