Improve nullif code generator
Originally, we cast the arguments of nullif to their common super types in code
generator. This is incorrect, they should be casted to the corresponding types
expected by the equals operator. Right now, since we don't have any equals
operator where the arguments are different, there is no perceivable difference
between the current and correct behavior.

This fix also makes it possible to remove a hack.