Rename property `query.max-age` to `query.min-expire-age`.
This property old name is missleading, suggesting that something will
definitly happen to query after reaching `query.max-age`.
The truth is that the value desribes the minimal time after which the
query MAY but musn't be dropped from history.
So this time is minimum required in order something (query drop) to happen,
yet reaching this age is not enough expire the query.

More information about query expiration policy may be found in
SqlQueryManager::removeExpiredQueries method.

WARNING: This change may invalidate old configurations!