parseTypeSignature automatically detects expressions
This commits extends type signature parser to handle
literal computations which may be passed as arguments.
Currently we need to determine the types of all tokens just after
parsing type signature.
Doing proper semantic analysis would be better solution but would
require a lot of changes to the codebase.

For now we assume that any token which contains arithmetic ('+', '-',
'/', '*') or is call min() or max() function is not a type parameter but
literal calculation parameter.

This is just intermediate step. Later we change codebase so literal
calculations are not allowed directly as parameters to parsed type signatures.
We require the calculation  be first bound to variables and stored in function context.
And only allow bound variables as literal parameters.