Fix `hive.orc.stream-buffer-size` property usage.
The property was unused but it seems like it was intended to be used
as default of `orc_stream_buffer_size` session property in the same
manner as the `hive.orc.max-buffer-size` is used as default for
`orc_max_buffer_size`.

This looks like typo in 1376164ce3de09c555d9e370bbacf5cf293fe9a5.