Allow time and timestamp without time zone in AT TIME ZONE
This commit fixed the bugs in ExpressionAnalyzer and DesugaringRewriter:
1. AT TIME ZONE should allow time and timestamp with or without time
zone as its input while the old ExpressionAnalyzer only allows time and
timestamp with time zone.
2. when rewriting AtTimeZone, value with time must be cast to time with
time zone, and value with timestamp must be cast to timestamp with time
zone.

Since AtTimeZone is being rewritten after planning, AtTimeZone could not
be compiled and executed without planning. Move the AtTimeZone tests
from TestDateTimeFunctions to AbstractTestQueries in which the queries
could be exeucted by the full engine.