Fix potential deadlock in phased scheduler
The code was incorrectly ignoring the fragment that contains
the TableScan or Values nodes as a source and missed an edge
in the scheduling dependency graph if there was a broadcast
join (probe + join in in the same fragment).

This could cause the join-probe stage to be scheduled independently
(and before) the build stage in a contented cluster, which
would result in the query deadlocking.