Add check that integral types coerce to BIGINT
Without this check, integral types narrower than BIGINT (ie INTEGER,
SMALLINT, and TINYINT) would be coerced to DOUBLE instead of BIGINT
if the DOUBLE variant of the function happened to be added to the
function list first.

This also required changes to TypeRegistry to deal with parametric
types with no parameters (ie VARCHAR).