Consider queued pages when checking if a buffer is finished
When a client fetches data from a SharedBuffer partition, the result contains
a finished state for the buffer if no pages were fetched and no more pages
can be added to the shared buffer. This is wrong because pages queued to be
added to the partiton are not taken into account.  This can result in lost
pages.