Improve performance of CLI output
The purpose of AnsiConsole is to install an AnsiOutputStream which
either strips out ANSI escape sequences or interprets them differently
on Windows. The CLI only uses ANSI in interactive mode, so we don't
need to pay the substantial performance overhead of filtering the output
in batch mode (such as when redirecting the output to a file).