Remove usages of query stats for verifying number of read rows
Fixes #4964. Fixes #5019.

This is a workaround as query stats should be used in such tests to
verify how much data was read to see that partition was pruned from
reading. However, due the fact that currently query stats are not
reliable and may return not up-to-date data it was changed to avoid
intermittent failures.

Now partition pruning is checked in a following way. Two partitions are
loaded. One correct and one with incorrect data. When a SELECT is done
only from correct partition, it is passing as partition prunning has
worked. When SELECT reads all data it fails (or return missing data for text)
due the read from incorrect partition.

This commit changes only TestTablePartitioningSelect.
TestTablePartitioningInsertInto is still using QueryStats. It is not
changed as it is under quarantine. As this solution is a workaround, and
using query stats is a desired behaviour I left
TestTablePartitioningInsertInto not changed.