Fix hashcode operator for varbinary
Due to legacy reasons, we have code for computing the hash code
of a type in two places: hash_code operator and block.

The update to 64-bit hashes missed a change to varbinary's
HASH_CODE operator, so the operator and the block produced
different hashes for the same data.

Eventually, we should converge to having just the hash_code
operator implementations and get rid of the block methods.