Change integral literals to fit in narrowest type
Integral literals will now be resolved as the narrowest possible
integral type. For example, the literal '1' will now be resolved to
the integer type, and the literal '10000000000' will be resolved to
the bigint type. This means that literals that are smaller than
INT_MAX but that are required to be BIGINTs will be represented as
GenericLiteral("BIGINT", number). This is a natural consequence of the
fact that we do not attach type information to the intermediate
representation of the expressions in the plan.