Add ROW constructor
It doesn't explicitly set the fields names.
Now it's possible to do nice things like `select row(1, row(9, a, array[], null), row(1, 2)), row(date '2016-04-15') from (values ('a'), ('b')) t(a);`

Notice that the type for ROW will have changes.
Now it's syntax is `ROW<type1, type2...>(field1, field2...)`, but it should be `ROW(field1 type1, field2 type2...)`
Because of the presence of `<` and `>` in the type name, this commit introduces a hack because the typeName is used as an part of variables name in the generated bytecode. Once it's changed to the new syntax, we can remove the hack.

Also, ROW CAST is not completly implemented:
- CAST from one row to another doesn't rename fields.
- It doesn't CAST the arguments types to super types.
- It's not possible to cast from NULL (implicitly) because it generates a type with fields with uppercase letters.
- It's not possible to explicitly cast because the the file `SqlBase.g4` doesn't define the syntax for ROW type.
- Because of the same reason you can get a weird error if you try `select * from (values (row(1,2), 3), (null, 3));`. Since arilift doesn't have the syntax for `ROW<types>(fields)` it throws `ParsingException`.