Support Hive Varchar for CREATE TABLE and INSERT
Before this path, Presto's `VARCHAR` type with any length was translated
to Hive's `STRING` type. Hive supports `VARCHAR` column with the length
less than or equal to `65535`. So the Presto's `VARCHAR` type with the length
less or equal than `65535` should be translated to the Hive `VARCHAR` type
with the same length. VARCHAR(MAX_SIZE) type should be  translated to the
`STRING` Hive type as it is now. VARCHAR columns with the length
less than `MAX_SIZE(0x7fffffff)` and greater than `65535` are not supported
by Hive connector. Upon attempt to create table in Hive
connector with such columns `Unsupported type` exception is being
thrown.