Implement PREPARE
Prepare queries execute in much the same way as set session queries.
Because Presto does not maintain any session state, updates to the list
of prepared statements are sent via http headers, and the list of
prepared statements is sent with each query from the client, also by
http header.