Add product test for GRANT and REVOKE
The product test for GRANT and REVOKE will be run under a single profile,
namely singlenode-kerberos-hdfs-impersonation. Running the product test
requires one to set "hive.security=sql-standard" in hive.properties file.
Setting this property turns on SQL Standard authorization for Hive
connector. We are restricting this change to a single profile so that only
the product tests in this profile will be run with SQL Standard
authorization turned on.