Move Signature creation outside of SqlFunction.
This refactor is first step to creating Signature of function ASAP
instead of passing it as fragmented fields throughout the code.
This will allow to simpler code flow, smaller number of argument
and more consistent interfaces.

Other focus point in this commit is to avoid passing Types as strings.
The general idea is to get to the state where type will be converted
from String to TypeSignature ASAP (while parsing annoations).
For now it just moves String types outside of SqlScalarFunction.

Future steps is to apply this changes to other SqlFunction based
classes. And to remove Signature constructors accepting Strings instead
of TypeSignature.