Fix creating Hive tables on S3 in empty database
The Hive connector has a sanity check that validates the existence
of the database location. This check fails on S3 because S3 only
has files, not directories, so a directory will not exist unless
one of the child directories has files.

A simple fix is to skip the check when operating on S3.