Capture partition count in Plan
When local parallel is enabled, the coordinator creates a query plan that
requires parallel execution, but parallelism is set on each worker using
the task concurrency property.  If the coordinator and worker have different
settings for this property, the execution can fail.

Instead of dividing the parallelism planning between the coordinator and the
worker, the coordinator makes the decision and records this in the plan.