Add ExecutorService for metric retrieval
This service will add parallelism to the retrieval of metrics
from the index metrics table, decreasing the overall planning time.
A new property, index_short_circuit_cardinality_fetch, is used to enable
a new feature which causes the metrics retrieval to return early if one
of the cardinalities retrieved is less than a given threshold.  The
remaining tasks will complete as planned, caching the results in the
Guava cache, but the connector can use the smallest cardinality
column in the meantime.

We also add a polling duration because some cached results are
returned a few milliseconds before a significantly lower result.
By adding a sleep prior to polling the tasks, we are adding 'waves'
of result retrieval.  The results of any completed tasks are taken
and the smallest cardinality, if below the threshold, is used while
the other tasks complete.