Fix grouping on aggregation arguments
Modify the GroupIdNode and GroupIdOperator to pass aggregation
arguments through when they are also grouped on. Previously, the
aggregation arguments and grouping columns would reference the same
symbol, which leads to unintuitive behavior. The problem only occurred
when there were multiple grouping sets, only some of which contained
the column that happens to be the aggregation argument.