Fix grouping on aggregation arguments
- Modify the GroupIdNode and GroupIdOperator to pass aggregation
arguments through.
- Change PruneUnreferencedOutputs optimizer to only
pass through columns that are also aggregation arguments.

The expected behavior is that arguments to the aggregations reference
values in the source relation rather than the outputs of the GroupId
operation (which can produce nulls for these columns). Prior to this
change, aggregation arguments and grouping columns referenced the same
symbol, producing incorrect results.

# Conflicts:
#	presto-main/src/main/java/com/facebook/presto/operator/GroupIdOperator.java
#	presto-main/src/main/java/com/facebook/presto/sql/planner/QueryPlanner.java
#	presto-main/src/test/java/com/facebook/presto/operator/TestGroupIdOperator.java