Fix the failure when chaining AT TIME ZONE
Chaining `AT TIME ZONE` results in the value part of the outer `AT TIME
ZONE` being rewritten, while the map of types is not updated, then
`getType` with the rewritten value expression will return a `null`,
which causes the query failure.

In this commit, we invoke `getType` the old value part of the outer `AT
TIME ZONE`, which should fix the problem without updating the map of
types.