Build docs, rpm and checks once per build
This cuts at least 100 s on the `install:` step on Travis.

The commit moves parts of build previously executed
for each matrix row to the end of `INTEGRATION_TESTS`
 branch. The mentioned parts are:
  - `presto-docs`
  - `presto-server-rpm`
  - dependency checks
  - style checks
  - all other airbase checks enabled by default

Those parts are moved *after* integration tests because
if the checks fail, it's useful to know that the tests pass
and that only code style / dependencies have to be fixed.