Introduce SymbolReference expression node
Until now relation object reference and symbol reference were modeled in
expression tree as QualifiedNameReference. This was misleading as these
two references points to different domain AST or logical plan.
They were used interchangeably and so it was not clear to which domain
given reference points. To avoid that SymbolReference get introduced.
Now these two domains are separated.