Add a row-based threshold for the smallest cardinality lookup
This new session property adds the ability to have a hard threshold for
the cardinality lookup in addition to the percentage-based threshold.
The actual value used for the early return threshold is the minimum
value of the percentage-based threshold multiplied by the number of
rows and the value of the new row-based threshold.