Update correlation in Apply node
All the column references of subquery plan which are not resolved (do
not come from subquery relations) are collected. SubqueryPlanner check
if it is able to rewrite such references to symbol produced by its
relation, if it is able then add replace column reference with that
symbol and that this symbol to the correlation list of Apply node. If
SubqueryPlanner is not able to rewrite that column reference, then it is
assumed the outer SubqueryPlanner execution will do that. If there is no
outer SubqueryPlanner then SemanticException should be thrown.