Do not ignore null intermediate aggregation states
`AccumulatorStateSerializer.deserialize` was only called for non-null positions
in the block. The `combine` function for aggregation function was also ignored.
This was an unintended side effect of the original AccumulatorCompiler
implementation. As a result, the `otherState` in `combine` function would be
guaranteed to not have any `null` fields if the
`AccumulatorStateSerializer.deserialize` implementation for the aggregation
function never produce such `State` objects when the input parameter to
`deserialize` is not null, which is normally true.

With this guarantee, a few aggregation functions passed all tests yet their
implementation seemed wrong. It makes reasoning about correctness very hard.
Additionally, when some guarantees exist but are not clearly defined, new
implementation are prone to depend on false guarantees. Furthermore, arguably,
there may be valid use cases for using null as the serialized form for a
special case, which should not be ignored.