Test for non correlated scalar subquery as JOIN conditition
Previously there were tests which were check if it is possible to filter
with non correlated scalar subquery the results of join. Hovewer this is
not much different that fitlering the table scan, so such tests didn't
have much value.

Instead, I moved the non correlated scalar subquery to be used as join
condition.