Properly cancel query if the pager is aborted
When the pager is aborted the main thread gets interrupted and statement
client close() is called with the interrupt flag of the main thread set.
This causes an `java.nio.channels.ClosedByInterruptException` under the
hood when making the http call, which fails and leaves the query
running until it gets canceled due to inactivity.