Change access controls for view creation from other views
This changes the API for checkCanCreateViewWithSelectFromTable to no
longer consider the destination view name. This makes it consistent
with checkCanCreateViewWithSelectFromView, which also does not have
this information.

Access control plugin implementations that want to deny creating views
with specific names can continue to do so by denying access in
checkCanCreateView.