Add DataTypeTest and TestTable for data type round trip tests
===

DataTypeTest allows writing data round trip tests in an abstract
and duplication-free way. For example:

DataTypeTest.create()
        .addRoundTrip(varcharDataType(10), "John Doe")
        .addRoundTrip(varcharDataType(2), "US")
        .execute(queryRunner, dataSetup);

will:
- create a table and populate it with data using the dataSetup strategy,
- select all rows, check that there's exactly one row and that it has:
  - columns of types as specified,
  - values as populated during data setup,
- drop the table.

The provided DataSetup strategies are:
 - new CreateAsSelectDataSetup(sqlExecutor, tableNamePrefix),
 - new CreateAndInsertDataSetup(sqlExecutor, tableNamePrefix).

The SqlExecutor is a thin facade providing a single interface over both
Presto QueryRunner and JDBC drivers. The respective implementations are:
 - new PrestoSqlExecutor(queryRunner)
 - new JdbcSqlExecutor(jdbcUrl).

Note that a single instance of DataTypeTest can be execute-d multiple
times, e.g. with different dataSetup-s to test both create as select and
create + insert syntax on Presto (to check it's correctness). The same test
would then be used with the target database's sqlExecutor to check that
Presto inserts the data in the same way as the target DB does.

===

The TestTable is a simple AutoCloseable for automatic creation
and disposal of tables. Its constructor is:

new TestTable(sqlExecutor, tableNamePrefix, ddlTemplate)

The constructor will use the sqlExecutor to create a table using ddl
resulting from replacing '{TABLE_NAME}' in the ddlTemplate with
a unique name starting with tableNamePrefix.