Send a delete to the remote task when task transitions to done state
When a task transitions to a done state, we might not have the final
stats for the task. So send a delete to the remote task which will
return the final stats for the task. With this change we can update the
TaskInfoFetcher to remove the logic to send the last request.

Note: we already send a delete to the remote task from the
PageBufferClient to free up the output buffers. Since the delete
operation is idempotent, sending a delete again is harmless.