Fix JDBC metadata for floating point
Return proper values for COLUMN_SIZE and NUM_PREC_RADIX in case where a
type is REAL or DOUBLE. In Presto REAL and DOUBLE are IEEE 754
single/double-precision floating number and precisions are 53 and 24 bits.

This also changes the value of NUM_PREC_RADIX in the case where a type
is a number. It no longer returns 10 for non-number types like varchar
or date but instead returns NULL.