Remove WindowNodes that have no window functions
PruneUnreferencedOutputs will removed window functions from window nodes that
are not referenced elsewhere. This can result in a window node that has no
window functions at all. We need to drop the window node from the plan so as
not to repartition the data when we aren't going to do any computation with it.