Prune unreferenced Apply nodes from plan
When a subquery result does not affect the result of a whole query it
can be pruned. In order to achieve that when Expression which is using
subquery (Apply) results is pruned we need to remove matching Apply
node.

This patch removes only Apply nodes related to expressions, like:
... 'ala' IN (SELECT name FROM names) ...
... EXISTS (SELECT name FROM names) ...
... (SELECT name FROM names) ...

Currently it prunes Apply for expressions from unused projections, like:
SELECT count(*) FROM (SELECT <UNUSED_PROJECTION> FROM ...) ...