Test for non correlated EXISTS subquery as JOIN conditition
Previously, tests for EXISTS with join were checking if it is possible to filter
with non correlated EXISTS subquery the results of join. However this is
not much different that filtering the table scan, so such tests didn't
have much value.

Instead, I moved the non correlated EXISTS subquery to be used as join
condition.