Move test for non correlated scalar subquery to JOIN conditition
Previously, tests for join with scalar subquery check if it is possible to
filter with non correlated scalar subquery the results of join. However this
is not much different that filtering the table scan, so such tests didn't
have much value.

Instead, I moved the non correlated scalar subquery to be used as join
condition.