Add ReorderWindows optimizer
Adjacent windows of the same partition by clause should be next to each other in the plan,
to avoid unnecessary repartitionings.

What's more, whenever there are partitions where one is a prefix of another, the shorter comes first because of the same reasons.

This optimizer is limited to operate on adjacent windows only.
It does not reorder windows across other PlanNodes than WindowNode.