Fix aggregation over grouping sets to produce default output
When an aggregation with an empty grouping set, cube or rollup
processed no input rows, no output was produced. This is incorrect, as
empty grouping sets are considered to be global aggregations and must
generate a single row even if no input was processed.

This commit makes the following changes:

- Both AddExchanges and AddLocalExchanges detect empty grouping set,
  and collapse to a single final aggregration operator.
- HashAggregationOperator now generates default outputs when there are
  global aggregations and no input has been processed.
- Adds tests for several cases when no input is received.