Fix a leak of futures in DeleteOperator
Currently a new listenable future is created on every "isBlocked()" call
in DeleteOperator which I expect to happen often. The issue with it is
that a new listenable future is referenced by completable future and as
a result cannot be cleaned up by GC. Furthermore, every time it happens
a new callback is added to a completable future which is not ideal.
The fix is to do conversion immediately when a future is created.