Don't deduplicate window functions in QueryPlanner
When two window functions existed in a query and one was coercing the other didn't get a symbol assigned.
E.g. in
  select 1.0*row_number() over(), row_number() over()
QueryPlanner was building plan based on one function only (because both row_number calls are identical and were deduplicated).
While building a plan the coercion was triggered for the first window function though.
It caused creating a symbol referring to an expression CAST(row_number() AS DOUBLE)
which is not reusable by second window function anymore.

Deduplication is necessary to handle case when OrderBy clause refers to window function from outputs though.