Add rule to evaluate a constant Apply expression
Changing the order of the optimizers in 4791a802a69f5f73d21c43b11df9b5980879f97c
causes some plans to end up with an Apply node at the end of the optimization.
Apply is not supported at execution time, so queries of that form fail with
a plan validation error.

This change introduces a new rule to reduce subplans of the following shape:

- Apply[]
  - Values (input)
  - Values (subquery)

and turn them into a simple Values node.