Implement FILTER clause for aggregations
Implement the filtered aggregation syntax.  In this change, a
FilterClause is added to the FunctionCall class to have the
aggregate to have a boolean expression as a filter. The
expression will be then used as masks to mask the rows that
are not satisfy the expression.

Current implementation will not support both distinct and
filter due to the masks will over write each other.  Will
re-visit this scenario in future development.