Build other modules as a special case
Explicitly excluded module with '-pl !module' won't be included to
build even when the `-am` is specified, and some of the not excluded
modules rely on that.

The code of the maven graph builder is easy to follow and can be found here:
https://github.com/apache/maven/blob/master/maven-core/src/main/java/org/apache/maven/graph/DefaultGraphBuilder.java

For the job when modules are excluded we need to build the entire project. And than run tests on that module.