Build other modules as a special case
An explicitly excluded module with '-pl !module' won't be included
in the reactor even when '-am' is specified and thus is not
available for non-excluded modules that depend on it.

For the job that uses excluded modules, we need to build the entire
project and then run tests on that module.