Make Plan DSL independent of SymbolAllocator behavior
The plan DSL is currently dependent on the behavior of the SymbolAllocator;
creating aliases in tests requires knowing some substring of the symbol that
the SymbolAllocator generates.  This implicit dependency is brittle and is a
source for ambiguity that can lead to tests that pass incorrectly.

This change makes it so binding aliases in unit tests is done against the
rvalue that is being assigned to the symbol. This makes the DSL independent of
the SymbolAllocator's behavior.