Fix NPE when scheduling non-remotely accessible splits
When scheduling a large initial batch of non-remotely accessible splits,
the split queue could fill up before the RemoteTask was created. This
causes a NPE when attempting to get the queue space future from the
task.
Instead, we just skip these nodes as the RemoteTask will be created
immediately after split scheduling.