Fix excess memory usage in HttpRemoteTask
If an HttpRemoteTask ran for a long time without transitioning to a
state where more splits could be queued, and many callers added
dependent futures to the CompletableFuture returned from
whenSplitQueueHasSpace(), there could be a large amount of memory used by
those dependents because they aren't cleaned up.
Due to JDK-8161600 there is no way to fix this with CompletableFutures, so
the implementation has been changed to use a FutureStateChange.