Remove queryBoundary from Scope
queryBoundary was designed to point a scope at which query scope changes
(from inner to outer). It appeared pointless as when a scope points at
the same query nesting level it contains all the query entities. So when
a field is accessed from the other than current scope then it is
external (no jump over queryBoundary is needed).