Add workaround for CassandraQueryRunner tpch schema overwrite
By analogy to AccumuloQueryRunner

TestNG instantiates all the test classes on the startup and accepts
query runners as a constructor parameters

TestCassandraDistributed needs the entire TPCH schema to be provisioned,
when TestCassandraIntegrationSmokeTest needs only ORDERS table.

If TestCassandraIntegrationSmokeTest loads after the TestCassandraDistributed,
the schema with all TPCH tables being overwritten with TPCH schema with only
one single table.

This is only temporary workaround. As a long term workaround the way QueryRunner
is being passed to test framework must be changed. Say, it must be passed via
Supplier, and instantiated only when TestNG enters the test class.