Fix isBlocked() behaviour in TableWriterOperator
Operator#finish() is called when source operator is completed. So, it may be called
before Operator#addInput() is finished on the destination operator. In such situation
the destination operator must still be in a BLOCKED state, until the addInput() operation
is completed.

Instead of overwriting the addInput() future with the finish() future it is needed to
keep them both. isBlocked() must return one that is not yet completed.