Add support for date/time types in JDBC, fix some unit test quirks
When reading table metadata, if JDBC's metadata reports a 'timestamp' or
'time' column, this code checks to see if the database type name in the
metadata is 'timestamptz' or 'timetz' and advertises 'with time zone'
types as appropriate.

Times are stored in JDBC as an offset in milliseconds from 1970-01-01T00:00:00. For
times without timezones, they're interpreted in the timezone of the JVM.
For example, if the JVM's timezone is Katmandu (GMT+05:45), a local time
in milliseconds is offset by -345 minutes before being stored. For times
with timezones, the timezone packed with the time is used as the offset
before storage. The offset is always guaranteed to be positive.

Here's a worked example for a time without a timezone and the JVM in
Katmandu:

* String: '02:30'
* Milliseconds (long): 9000000
* Minus timezone offset of 345 minutes (java.sql.Time): -11700000ms
* Stored using JDBC `setTime`
* Retrieved using JDBC `getTime` (java.sql.Time): -11700000ms
* Made positive (java.sql.Time): 95400000ms

Timestamps are stored the same way (without trying to make them
positive, of course).

PostgreSQL does not store timestamps, even for `TIMESTAMP WITH TIME
ZONE`. When retrieving times/timestamps with time zones from PostgreSQL,
the UTC zone is always packed alongside the raw time.