Fix planning issue when ORDER BY clause contains aggregations
A recent commit (ec2e8976ba9cd59f3990b8e46288b8a728065b07) changed
the way ORDER BY expressions are handled in a way that causes certain
expression to not be "analyzed" and their types be recorded in the
Analysis object.

extractAggregates() looks for aggregations in node.getOrderBy()
and records them for later use by the planner. The new ORDER BY
analyzer process the rewritten expressions, which have a different
object identity. As a result, the aggregates don't have associated
type and implicit coercion information for the planner to use.

This change makes it so that the aggregates are extracted from
the rewritten expressions.