Fix reporting for errors in ORDER BY expression
Due to ec2e8976ba9cd59f3990b8e46288b8a728065b07, when analysis fails for certain
expressions, the error is misreported as happening in the SELECT clause
instead of in the ORDER BY clause. This is because the analyzer processes
the rewritten expressions, which contain inlined SELECT expressions and their
original locations.

This change fixes the issue by analyzing the original unmodified expressions
with a synthetic scope built from the output of the SELECT clause
that can delegate resolution to the source scope for missing names (essentially,
it implements the resolution rules per the SQL spec).

One side-effect of this change is that queries whose ORDER BY clause reference
columns that appear multiple times in the SELECT clause are now considered
invalid due to ambiguous references -- this matches the expected behavior
according to the ANSI spec.