Preserve relative order of conjuncts
ExpressionUtils.and() and ExpressionUtils.combineConjuncts()
try to build a balanced tree of terms. Due to the way the code
was written, for odd-numbered expressions, the algorithm places
the last term as the first term in the tree. For example:

and(a, b, c, d, e) -> and(e, and(and(a, b), and(c, d)))

This can cause the predicate pushdown optimizer to reorder terms
that are sensitive to evaluation order.

Additionally, combineConjuncts attempts to remove duplicate terms,
but it does not preserve the relative ordering of the terms in the
input.