Fix NPE when Like node has no escape
This was a regression introduced by 24407ce9de51c8d567990648d30d57e86e8a4e68
and a8d0940df0e777adf8b7bcd7c1e34546c7ffbeb2.

LikePredicate.getNodes() was not handling the case where the escape is
missing (null), and it was trying to add a null to an ImmutableList.