Fix quantified comparisons corner cases for <, <=, >, >= operators 
Previously the query rewrite for:

SELECT $VALUE {<, <=, >, >=} {ALL, ANY} ($SUBQUERY)

did not account for nulls in neither $VALUE nor $SUBQUERY. This
was giving erroneous results, namely:
  - NULL instead of TRUE/FALSE for $VALUE = NULL and
    a $SUBQUERY returning an empty relation
  - TRUE/FALSE instead of NULL for some queries for non-NULL
    $VALUE and a $SUBQUERY returning both NULL and non-NULL
    values.

This commit changes the rewrite, by making it compute - besides
the min/max value from the $SUBQUERY - also the counts of non-null
and all values returned by the $SUBQUERY and altering the full query
result to take nulls into account.

The new rewrite explicitly tests for empty $SUBQUERY result and returns
TRUE/FALSE for ALL/ANY accordingly. It also checks if there are NULLs in
$SUBQUERY results and adds a NULL conjunct/disjunct accordingly so
that the whole query evaluates to NULL.

This commit also adds numerous tests for the corner cases, comparing
with H2 for reference. I also checked that it's consistent with PostgreSQL.

 Adresses #6653.