Produce Hive splits for bucketed tables in round-robin fashion
This reduces the likelihood that the scheduler gets blocked when one worker has
more splits queued than limit while other workers have no splits. Without round
robin, for a bucketed partition, the split loader would produce a series of
splits that has node afinity like C, C, ..., C, A, A, ..., A, D, D, ..., D, B,
B, ..., B. If there are more splits for node C than the number of queued splits
allowed, node A, B, D would not have any split available to run because the
scheduler is blocked.

In addition, this commit changes the policy to determine the target size for
initial splits. The original policy tries to produce a number of initial splits
that have similar size. For example, assuming maxInitialSplitSize = 9K,
maxSplitSize = 30K, and the file size is 48K.

* When maxInitialSplits = 10, the file will be split into 8K, 8K, 8K, 8K, 8K, 8K.
* When maxInitialSplits = 2, the file will be split into 8K, 8K, 16K, 16K.

In the new policy,

* When maxInitialSplits = 10, the file will be split into 9K, 9K, 9K, 9K, 9K, 3K.
* When maxInitialSplits = 2, the file will be split into 9K, 9K, 30K.

You can see that the old policy is better for case 1, while the new policy is
better for case 2. A smart policy that are optimal in both cases are available.
However, such policy has to know exactly how many initial splits are left.
This is not possible given the parallel nature of BackgroundHiveSplitLoader.