Fix incorrect pruning of join output columns
Some nodes are sensitive to which columns are produced by
their children (e.g., DistinctLimitNode). The change to
add support for pruning join columns did not correctly
deal with this, which results in queries like this one
to produce incorrect results:

    SELECT DISTINCT x
    FROM (VALUES 1) t(x) JOIN (VALUES 10, 20) u(a) ON t.x < u.a
    LIMIT 100

Additionally, the constructor in JoinNode that derives
the output columns from the children is error-prone, since
it lends to misuse in places that expect the outputs not to
 change.

This change fixes the issue by making the specification of
join output columns explicit in every place that constructs
a Join node.