Process expression instead of group reference
When the result of applying a rule is a GroupReference
(e.g., in the case where a node is being removed, such as
for X->Y->(1) ==> X->(1)), make sure the loop continues
processing the expression contained in the target group.

Currently, the loop would try to continue processing
the GroupReference, which is non-sensical, since rules
can't match it.