Replace QualifiedNameReference with Identifier
This change introduces a new AST node, Identifier, designed to
represent an identifier atom. Qualified name references are now
encoded as DereferenceExpression across the board.

It does away with the special distinction between QualifiedNameReference
and DereferenceExpression. This is the first step in being able
to support quoted identifiers and other simplifications to the
analyzer.