Reintroduce queryBoundary in Scope
queryBoundary flag specifies if the parent scope
is stll part of same query or is external to the query
(e.g: parent scope of subquery). This is useful
for recognising if a field should have a corresponding
symbol in TranslationMap or if a field is external
to the query (e.g: correlated subquery field).