Don't merge windows when one depends on another
Windows won't be collected once a dependency is found.
Instead, they are dumped in the plan and collection starts over.

This may result in suboptimal plan but at least will be correct.
E.g. of suboptimal solution:
There are given windows A, B, C, D.
A and C have a specification X, B and D - specification Y.
Window A depends on an output of window B, C depends on an output of D.
For such a setup the sequence of: A -> B -> C -> D could produce an optimal plan of AC -> BD.
While with this patch it will result in A -> B -> C -> D (nothing merged) instead.
This is because after seeing A, the dependent B is next, causing to dump "stashed" A.
After B, C comes but they have different specifications, therefore - both are stashed for future Window Nodes to be merged with.
Then D appears in plan which is a dependency for already "stashed" C which causes dumping the stash.
Then, the plan looks like mentioned above.

If the order in plan was different though, e.g.: A -> C -> B -> D then the final plan would be: AC -> BD (both merges possible).

This is greedy approach but solves an issue with merging dependent windows (#7247).