Successful metadata queries not logged
Two state machines are used in class QueryStateMachine: queryState and finalQueryInfo. The query log
callback function was previously hooked only with finalQueryInfo, which does not have information with
the query state (RUNNING, FINISHING, etc). The state of finalQueryInfo will be updated only when the
output result is handled. This is not true for metadata queries, which do not have outputs. This
patch hooks the metadata query callback function with queryState. The callback is invoked when the
query finishes.