Disallow usages of IdentityHashMap in sql.{planner,analyzer} packages
IdentityHashMap is - due to its non-deterministic iteration order -
a potential source of query plan non-determinism. Therefore, a static
analysis check is added to prevent its usages in planner and analyzer.

The check is implemented using checkstyle instead of
maven-modernizer-plugin, because the latter doesn't offer a way to
restrict a violation rule to certain packages.