Make RuleAssert to check how many times rule can be applied
In most case rule can be applied only once, in some case it is expected
that rule can be applied several times. This change makes possible
developer to check how many times given rule can be applied to given
plan.

Thanks to that it will be easier to detect indefinitely recursive rules.