Replace CompletableFuture with ListenableFuture
The CompletableFuture interface is difficult to use correctly, and it is
easy to introduce subtle bugs that result in lost notifications. For example
CompletableFuture does not propagate cancel from wrapped futures to the
core future. Additionally, it is easy to accidentally schedule tasks on a system
wide fork join pool.

ListenableFuture on the other hand is easy to understand and reasonable, and
the transformations in Guava handle cancelation properly.