Refactor AbstractTestQueryFramework
Add constructor that accepts Supplier<QueryRunner>.
Deprecate the constructor that accepts QueryRunner.

TestNG initializes all the test classes at startup. When a QueryRunner
is passed as a constructor parameter, it is created at TestNG startup.
If there are many classes that extend AbstractTestQueryFramework in single
suite, that means that at startup as many QueryRunners will be created.