Fix handling of DATE types in AccumuloRecordCursor#getLong
The connector is incorrectly converting the number of days into a Date
object as if the number of days was milliseconds.  This works from a
data perspective, but doesn't align with the intention of using familiar
Java types.  This commit updates the data to use milliseconds while
maintaining backwards compatability with the existing serializers.