Ensure FileSystem is created with PrestoHadoopConfiguration
We use PrestoHadoopConfiguration to pass already created objects to
FileSystem instances. This custom Configuration class is created by
HdfsConfiguration.getConfiguration(), but not all code paths that can
create a FileSystem instance use this. We can hack around this by
creating the FileSystem eagerly and relying on the Hadoop FileSystem
cache that ensures all FileSystems instances are singletons.