Do not re-create Cassandra Cluster
This object is designed to be singleton.
Reconstructing it causes additional spawn of thread pools and connection pools.

The mechanism of re-creation of driver was introduced when driver version was 2.0.0-rc2,
to avoid problem when Cassandra cluster nodes configuration has changed.

With newer version of driver we no longer need this workaround.