Fix join property derivations when not all columns are output
Property derivation logic was producing properties in terms of
columns that might not be output by the join. This is an invalid
derivation.

As a result, AddExchanges could make suboptimal decisions when
deciding whether an operation could leverage the properties from
its inputs.

This change fixes it the logic to either remove or translate the
columns (where appropriate, based on equality conditions) so that
properties only reference output columns from the join.