Fix LDAP product tests for newer versions of docker compose
Newer versions of docker-compose fails for invalid host ":" when those
variables are not set. These variables are only being used when it is
required to substitute ldap server with the external one, say,
ActiveDirectory, which cannot be deployed with docker.

As a solution the default dummy values has been set for those
variables just to do not leave them empty.