Replace xargs with a loop in run_on_docker
When there are no containers to remove - xargs fails with a weird
message that there are argument is missing for "docker rm".

Linux version of xargs has -r arguments, which tells to do not
run a command when there are no arguments.

However xargs on Mac doesn't have such flag.