Make it possible to choose regex library on a per-query basis
Due to the complexity involved, when RE2J was introduced into
Presto as a regular expression library, we decided against having
a session property that swtiches between the two libraries.

This config makes the config property control which implicit cast
to turn on (instead of which set of functions to register). This
way, it is possible to choose the regex library by using an
explicit cast in the query.