Fix leak of query and task when cleanup request is rejected
When HTTP request sent by the coordinator to the worker to fail a task is
rejected by HttpClient with RejectedExecutionException, proper cleanup is not
carried out. The task is not marked as failed on the coordinator. Now that the
task is still marked as running on the coordinator, the coordinator will
continue to fetch task info from worker. Since HTTP request to fail the task
was rejected, the worker also considers the task still running because it never
got the message to fail the task.

As a result, the coordinator will continue to try to fetch task info from
workers because it's not in a done state. And the worker will report the tasks
as running when task info is requested. As a result, given both the coordinator
and the worker think the task is still running, the coordinator will continue
to continuously fetch the TaskInfo from the workers, keeping them from
expiration.

Notably, there are two places on the coordinator that keeps track of task
state. One in ContinuousTaskStatusFetcher. The other in TaskInfoFetcher. In
this particular case, the former correctly marked the task as failed, but the
latter didn't.