Add new interface to Block to get a breakdown retained sizes
BlockBigArray can set two blocks that share the same underlying data
with different indices. This leads to overcounting in some aggregation
functions such as max_by, min_by, or arbitrary. This patch adds a new
interface to Block to return the sizes for different data so that
blockBigArray can track different underlying data.