Change StatsCalculator API to use Lookup for computing child costs
Instead of recursing in the StatsCalculator api, pass the Lookup instance into
the stats computation. Then the method computing a stats for plan node
can use Lookup to compute stats (and possibly other traits) for source
nodes.

The StatelessLookup is a temporary measure for places that don't have
access to the lookup for individual plans. It can be used across
multiple queries because it doesn't keep state, but it won't resolve
GroupReferences.