Verify nonReserved contains only tokens
The mechanism for handling non-reserved keywords relies on being able to
intercept the parsing operation and replacing the nonReserved entry in the
parse tree with an IDENT token.

Modifying the tree in this manner breaks when dealing with intermediate
nodes, presumably due to how enter/exit rule notification is weaved into
the traversal/building of the tree.