Add PruneSemiJoinFilteringSourceColumns
Add and enable an iterative optimizer rule to introduce a project-off
above the filtering source when possible.  Separate from
PruneSemiJoinColumns because the needed columns from the filtering
source are independent of what is needed by the parent of the semijoin,
so we can't rely on a project being present above the semijoin to
trigger the rule.