Fix correctness bug in PageFunctionCompiler
Block builder was only being reset at the end of the projection, so if
an exception is thrown in processing, block builder contains data from
previous failed run. The DictionaryAwarePageFilter attempts to apply
projection to entire dictionary, and if this fails, the page is processed
as normal using the same projection function, which triggers this bug.