Add PruneTopNColumns Rule
Migrate PruneUnreferencedOutputs handling of TopNNode to the iterative
optimizer.  This rule finds a TopNNode that's under a project, such that
the grandchild produces columns not needed by either the parent or the
child, and creates a new project grandchild to discard the unused
columns.  Subsequent rules may match the pattern of the new grandchild
(project) over the old grandchild, and do further pruning.